issue/630 - slightly improved unimplemented messages
